requirements.txt
# Core
MetaTrader5==5.0.45
python-dotenv==1.0.0
pyyaml==6.0.1

# Data Analysis
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
# pandas-ta: Opcional, mas não disponível no PyPI padrão

# NLP & Sentiment Analysis
textblob==0.17.1
nltk==3.8.1

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
xgboost==2.0.3
lightgbm==4.1.0

# APIs & Web
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Telegram
python-telegram-bot==20.7
deep-translator==1.11.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Task Scheduling
APScheduler==3.10.4
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utils
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
yfinance>=0.2.28

# Technical Analysis  
# TA-Lib==0.4.28  # ⚠️ COMENTADO: Requer Visual C++ Build Tools (não usado no código)
